#158 [REFACTOR] URL 함수 분리 및 OnboardingGuard 생성
