+++ src/layouts/PublicLayout.tsx
@@ -1,59 +1,11 @@
-import { Outlet, useLocation, Navigate, useNavigationType } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
 import { ErrorBoundary } from 'react-error-boundary';
 import { Suspense } from 'react';
 import Loading from '../pages/Loading.tsx';
 import ServerError from '../pages/ServerError.tsx';
-import { useLocalStorage } from '../hooks/useLocalStorage';
-import { LOCAL_STORAGE_KEY } from '../constants/key';
-import onboardingSteps from '../constants/onboardingSteps';
 
 const PublicLayout = () => {
-  const location = useLocation();
-  const navType = useNavigationType();
-  const { getItem: getInviteUrl } = useLocalStorage(LOCAL_STORAGE_KEY.inviteUrl);
-  const { getItem: getInvitePassword } = useLocalStorage(LOCAL_STORAGE_KEY.invitePassword);
-  const { getItem: getIsInvite } = useLocalStorage(LOCAL_STORAGE_KEY.isInvite);
-
-  const guardedPaths = onboardingSteps.slice(1); // 온보딩 경로 중 `/onboarding`(0단계) 제외, 1~3단계 경로만 보호
-  const needsCheck = guardedPaths.includes(location.pathname);
-  const isManualOrHistoryNav = location.key === 'default' || navType === 'POP';
-
-  // 온보딩 진행 상태 관련 로컬스토리지 값
-  const inviteUrl = needsCheck && isManualOrHistoryNav ? getInviteUrl() : null;
-  const invitePassword = needsCheck && isManualOrHistoryNav ? getInvitePassword() : null;
-  const isInvite = getIsInvite();
-  const ONBOARDING_STATUS_KEY = 'onboarding-status';
-  const hasOnboardingStatus =
-    needsCheck && isManualOrHistoryNav && localStorage.getItem(ONBOARDING_STATUS_KEY) !== null;
-
-  // 새로고침 감지
-  const navEntry = performance.getEntriesByType('navigation')[0] as
-    | PerformanceNavigationTiming
-    | undefined;
-  const isReload = navEntry?.type === 'reload';
-
-  // 주소창 직접 입력만 감지 (새로고침 제외)
-  const isManualNav = location.key === 'default' && !isReload;
-
-  // 유효한 초대 URL인지 체크
-  const isValidInviteUrl = (url?: string | null) => {
-    return !!url && url.includes('invite?token=');
-  };
-
-  // 직접 입력일 때만 가드 실행
-  if (
-    needsCheck &&
-    isManualNav &&
-    (!isValidInviteUrl(inviteUrl) ||
-      !inviteUrl ||
-      !invitePassword ||
-      isInvite ||
-      hasOnboardingStatus)
-  ) {
-    return <Navigate to="/onboarding" replace />;
-  }
-
   return (
     <QueryErrorResetBoundary>
       {({ reset }) => (
