+++ src/utils/validateInviteUrl.tsx
@@ -1,0 +1,12 @@
+// src/utils/validateInviteUrl.ts
+export const validateInviteUrl = (url?: string | null): boolean => {
+  if (!url) return false;
+  try {
+    const u = new URL(url, window.location.origin);
+    const hasInvitePath = u.pathname.includes('/invite');
+    const hasToken = new URLSearchParams(u.search).has('token');
+    return hasInvitePath && hasToken;
+  } catch {
+    return false;
+  }
+};
